Function liquidity::permissions::check

pub fn check(key: &str, user: &Option<User>) -> Result<(), PermissionError>

Check the user's permissions to ensure they are allowed to use the API function

Arguments

key - A string that holds the permission key required for access to this API

user - The user object that holds the permissions

Returns

True if the user has permission, failure::Error if the user doesn't have permission or isn't logged in

Example

use liquidity::{permissions::{self, PermissionError}};

// Make a mock user
let user = Some(User {
    id: "".to_string(),
    permissions: vec!["view:election".to_string()]
});

let valid = permissions::check("view:election", &user);
let invalid = permissions::check("create:election", &user);
let not_logged_in = permissions::check("create:election", &None);

assert_eq!(Ok(()), valid);
assert_eq!(Err(PermissionError::NotAllowed), invalid);
assert_eq!(Err(PermissionError::NotLoggedIn), not_logged_in);